requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
mpi4py==3.0.3
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
torch==1.9.0
torchvision==0.10.0
typing-extensions==3.10.0.0
